moved Share folder
from RadicalCollaboration folder to pages
