+++ src/pages/Share/MediaInfo.js
@@ -5,7 +5,7 @@ import Button from '@material-ui/core/Button';
 import Container from '@material-ui/core/Container';
 import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
-import Link from '../../../components/common/Link';
+import Link from '../../components/common/Link';
 
 const CtiLogoSection = () => {
   const classes = useStyles();
