+++ src/pages/Share/index.js
@@ -10,7 +10,7 @@ import CardContent from '@material-ui/core/CardContent';
 import CardMedia from '@material-ui/core/CardMedia';
 import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
-import { GenericHeaderSection } from '../../../components';
+import { GenericHeaderSection } from '../../components';
 import GitHubButton from 'react-github-btn';
 import { TwitterTimelineEmbed } from 'react-twitter-embed';
