+++ src/pages/Share/styles.js
